refactor(core)!: merge the request queue frontend into a single `RequestQueue` class
`RequestProvider`, `RequestQueueV1` and `RequestQueueV2` no longer differ in
behaviour — request coordination (locking, queue-head management) is now an
internal concern of the storage client — so they are merged into a single
concrete `RequestQueue` class:

- `RequestProvider` becomes `RequestQueue`; the `request_provider.ts` and
  `request_queue_v2.ts` modules are removed and the implementation lives in
  `request_queue.ts`.
- `fetchNextRequest`/`markRequestHandled`/`reclaimRequest`/`isEmpty` delegate to
  the slim client; the queue-head, locking, consistency and
  `recentlyHandledRequests` bookkeeping is gone.
- `isFinished` returns `false` while a background add batch is in flight,
  otherwise `client.isEmpty()`.
- `isEmpty()` reflects only pending requests (the next `fetchNextRequest()`
  would return `null`), preserving the crawler's task-scheduling contract.
- `getRequest` is keyed by `uniqueKey`.

The `RequestProvider`/`RequestQueueV1`/`RequestQueueV2` exports are removed; use
`RequestQueue` instead. `RequestProviderOptions` is renamed to
`RequestQueueOptions`.

Part of #3075.
