+++ packages/core/src/storages/request_loader.ts
@@ -2,7 +2,7 @@ import type { Dictionary } from '@crawlee/types';
 
 import type { Request } from '../request.js';
 import type { IRequestManager } from './request_manager.js';
-import type { RequestQueueOperationInfo } from './request_provider.js';
+import type { RequestQueueOperationInfo } from './request_queue.js';
 
 /**
  * An abstract interface defining a read-only stream of requests to crawl.
